move metrics functions into metrics.rs
