+++ src/lib.rs
@@ -3168,15 +3168,6 @@ impl GooseAttack {
         Ok(())
     }
 
-    // Update metrics showing how long the load test has been running.
-    fn update_duration(&mut self) {
-        if let Some(started) = self.started {
-            self.metrics.duration = started.elapsed().as_secs() as usize;
-        } else {
-            self.metrics.duration = 0;
-        }
-    }
-
     // Let the [`GooseAttack`](./struct.GooseAttack.html) run until the timer expires
     // (or the test is canceled), and then trigger a shut down.
     async fn monitor_attack(
@@ -3287,119 +3278,6 @@ impl GooseAttack {
                     true,
                 );
                 // No need to reset local metrics, the worker is exiting.
-            }
-        }
-
-        Ok(())
-    }
-
-    // If metrics are enabled, synchronize metrics from child threads to the parent. If
-    // flush is true all metrics will be received regardless of how long it takes. If
-    // flush is false, metrics will only be received for up to 400 ms before exiting to
-    // continue on the next call to this function.
-    async fn sync_metrics(
-        &mut self,
-        goose_attack_run_state: &mut GooseAttackRunState,
-        flush: bool,
-    ) -> Result<(), GooseError> {
-        if !self.configuration.no_metrics {
-            // Check if we're displaying running metrics.
-            if let Some(running_metrics) = self.configuration.running_metrics {
-                if self.attack_mode != AttackMode::Worker
-                    && util::timer_expired(
-                        goose_attack_run_state.running_metrics_timer,
-                        running_metrics,
-                    )
-                {
-                    goose_attack_run_state.running_metrics_timer = time::Instant::now();
-                    goose_attack_run_state.display_running_metrics = true;
-                }
-            }
-
-            // Load messages from user threads until the receiver queue is empty.
-            let received_message = self.receive_metrics(goose_attack_run_state, flush).await?;
-
-            // As worker, push metrics up to manager.
-            if self.attack_mode == AttackMode::Worker && received_message {
-                #[cfg(feature = "gaggle")]
-                {
-                    // Push metrics to manager process.
-                    if !worker::push_metrics_to_manager(
-                        &goose_attack_run_state.socket.clone().unwrap(),
-                        vec![
-                            GaggleMetrics::Requests(self.metrics.requests.clone()),
-                            GaggleMetrics::Tasks(self.metrics.tasks.clone()),
-                        ],
-                        true,
-                    ) {
-                        // GooseUserCommand::Exit received, cancel.
-                        goose_attack_run_state
-                            .canceled
-                            .store(true, Ordering::SeqCst);
-                    }
-                    // The manager has all our metrics, reset locally.
-                    self.metrics.requests = HashMap::new();
-                    self.metrics
-                        .initialize_task_metrics(&self.task_sets, &self.configuration);
-                }
-            }
-        }
-
-        // If enabled, display running metrics after sync
-        if goose_attack_run_state.display_running_metrics {
-            goose_attack_run_state.display_running_metrics = false;
-            self.update_duration();
-            self.metrics.print_running();
-        }
-
-        Ok(())
-    }
-
-    // When the [`GooseAttack`](./struct.GooseAttack.html) goes from the `Starting`
-    // phase to the `Running` phase, optionally flush metrics.
-    async fn reset_metrics(
-        &mut self,
-        goose_attack_run_state: &mut GooseAttackRunState,
-    ) -> Result<(), GooseError> {
-        // Flush metrics collected prior to all user threads running
-        if !goose_attack_run_state.all_users_spawned {
-            // Receive metrics before resetting them.
-            self.sync_metrics(goose_attack_run_state, true).await?;
-
-            goose_attack_run_state.all_users_spawned = true;
-            let users = self.configuration.users.clone().unwrap();
-            if !self.configuration.no_reset_metrics {
-                // Display the running metrics collected so far, before resetting them.
-                self.update_duration();
-                self.metrics.print_running();
-                // Reset running_metrics_timer.
-                goose_attack_run_state.running_metrics_timer = time::Instant::now();
-
-                if self.metrics.display_metrics {
-                    // Users is required here so unwrap() is safe.
-                    if self.metrics.users < users {
-                        println!(
-                            "{} of {} users hatched, timer expired, resetting metrics (disable with --no-reset-metrics).\n", self.metrics.users, users
-                        );
-                    } else {
-                        println!(
-                            "All {} users hatched, resetting metrics (disable with --no-reset-metrics).\n", users
-                        );
-                    }
-                }
-
-                self.metrics.requests = HashMap::new();
-                self.metrics
-                    .initialize_task_metrics(&self.task_sets, &self.configuration);
-                // Restart the timer now that all threads are launched.
-                self.started = Some(time::Instant::now());
-            } else if self.metrics.users < users {
-                println!(
-                    "{} of {} users hatched, timer expired.\n",
-                    self.metrics.users, users
-                );
-            } else {
-                println!("All {} users hatched.\n", self.metrics.users);
             }
         }
